requirements.txt
telethon
aiohttp
python-dotenv
aiogram
requests